Added HTML filtering to t_html_generator
Added the ability to prevent items from being exported by putting the
string NOHTML in the item’s comment.

Also to get this to build I needed to change all the instances of
code.google.com/p/gomock/gomock to github.com/golang/mock/gomock.  I
looked in the master thrift github and this is indeed the URL they are
using now.  See here:
https://github.com/apache/thrift/search?utf8=%E2%9C%93&q=gomock

I also updated README.md to chronicle what it took to get with rift to
build.